[legato-af] Check private header existence in do_install
root cause:
args.h have been removed from legato source

modification:
check if private header existence when performing do_install

Resolves: LXSWIREF-136

+++ common/recipes-legato/legato-af/legato-af_git.bb
@@ -65,7 +65,7 @@ do_install() {
     install -d ${D}/usr/share/legato/src
     cd ${S}/framework/c/src
     for file in eventLoop.h log.h args.h; do
-        install $file ${D}/usr/share/legato/src/
+        [ -e $file ] && install $file ${D}/usr/share/legato/src/
     done
 
     # liblegato
